1. Добавлены изменения в метод GetDataForCommonInformation в классе StatisticHelper: исправлены запросы получения данных, добавлено округление значений. 2. Изменены типы полей в классе CommonInformationModel. 3. Внесены изменения на страницу вывода заправок: убрана колонка "Забыл внести предыдущую заправку", добавлен значок вместо колонки.

+++ CarNotes/Models/CommonInformationModel.cs
@@ -10,11 +10,11 @@ namespace CarNotes.Models
         public CommonTimeOnSite TotalTime { get; set; }
         public int CommonMileage { get; set; }
         public int RefuelAmount { get; set; }
-        public int RefuelCost { get; set; }
+        public double RefuelCost { get; set; }
         public int RepairAmount { get; set; }
-        public int RepairCost { get; set; }
+        public double RepairCost { get; set; }
         public int ExpenseAmount { get; set; }
-        public int ExpenseCost { get; set; }
-        public int AverageFuelPrice { get; set;}
+        public double ExpenseCost { get; set; }
+        public double AverageFuelPrice { get; set;}
     }
 }
